remove useHistoryPagination; share updateURLParams
