+++ core/web/components/events/list.tsx
@@ -1,8 +1,7 @@
 import { useState } from "react";
 import { useApi } from "../../hooks/useApi";
-import { useOffset } from "../../hooks/useOffset";
+import { updateURLParams, useOffset } from "../../hooks/URLParams";
 import { useSecondaryEffect } from "../../hooks/useSecondaryEffect";
-import { useHistoryPagination } from "../../hooks/useHistoryPagination";
 import { useRealtimeModelStream } from "../../hooks/useRealtimeModelStream";
 import Link from "next/link";
 import { useRouter } from "next/router";
@@ -37,7 +36,6 @@ export default function EventsList(props) {
   const limit = 100;
   const { offset, setOffset } = useOffset();
   const [type, setType] = useState(router.query.type?.toString() || "");
-  useHistoryPagination(offset, "offset", setOffset);
 
   useSecondaryEffect(() => {
     load();
@@ -70,17 +68,7 @@ export default function EventsList(props) {
       }
     }
 
-    updateURLParams();
-  }
-
-  function updateURLParams() {
-    let url = `${window.location.pathname}?`;
-    if (offset && offset !== 0) url += `offset=${offset}&`;
-    if (type && type !== "") url += `type=${escape(type)}&`;
-
-    const routerMethod =
-      url === `${window.location.pathname}?` ? "replace" : "push";
-    router[routerMethod](router.route, url, { shallow: true });
+    updateURLParams(router, { offset, type });
   }
 
   async function autocompleteProfilePropertySearch(match?) {
