+++ core/web/components/export/list.tsx
@@ -1,8 +1,7 @@
 import { Fragment, useState } from "react";
 import { useApi } from "../../hooks/useApi";
-import { useOffset } from "../../hooks/useOffset";
+import { useOffset, updateURLParams } from "../../hooks/URLParams";
 import { useSecondaryEffect } from "../../hooks/useSecondaryEffect";
-import { useHistoryPagination } from "../../hooks/useHistoryPagination";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { Row, Col, ButtonGroup, Button, Badge, Alert } from "react-bootstrap";
@@ -23,7 +22,6 @@ export default function ExportsList(props) {
   const limit = 100;
   const { offset, setOffset } = useOffset();
   const [state, setState] = useState(router.query.state?.toString() || "");
-  useHistoryPagination(offset, "offset", setOffset);
 
   let profileGuid: string;
   let destinationGuid: string;
@@ -40,7 +38,7 @@ export default function ExportsList(props) {
   }, [offset, limit, state]);
 
   async function load() {
-    updateURLParams();
+    updateURLParams(router, { state, offset });
     setLoading(true);
     const response = await execApi("get", `/exports`, {
       limit,
@@ -77,16 +75,6 @@ export default function ExportsList(props) {
   function formatTimestamp(timestamp) {
     const [date, time] = new Date(timestamp).toLocaleString().split(",");
     return `${date} ${time}`;
-  }
-
-  function updateURLParams() {
-    let url = `${window.location.pathname}?`;
-    if (state) url += `state=${state}&`;
-    if (offset && offset !== 0) url += `offset=${offset}&`;
-
-    const routerMethod =
-      url === `${window.location.pathname}?` ? "replace" : "push";
-    router[routerMethod](router.route, url, { shallow: true });
   }
 
   return (
