+++ core/web/pages/apiKeys.tsx
@@ -2,10 +2,9 @@ import Head from "next/head";
 import { Button } from "react-bootstrap";
 import { useRouter } from "next/router";
 import { useApi } from "../hooks/useApi";
-import { useOffset } from "../hooks/useOffset";
+import { useOffset, updateURLParams } from "../hooks/URLParams";
 import { useState } from "react";
 import { useSecondaryEffect } from "../hooks/useSecondaryEffect";
-import { useHistoryPagination } from "../hooks/useHistoryPagination";
 import Link from "next/link";
 import Pagination from "../components/pagination";
 import LoadingTable from "../components/loadingTable";
@@ -24,14 +23,13 @@ export default function Page(props) {
   // pagination
   const limit = 100;
   const { offset, setOffset } = useOffset();
-  useHistoryPagination(offset, "offset", setOffset);
 
   useSecondaryEffect(() => {
     load();
   }, [limit, offset]);
 
   async function load() {
-    updateURLParams();
+    updateURLParams(router, { offset });
     setLoading(true);
     const response = await execApi("get", `/apiKeys`, {
       limit,
@@ -42,15 +40,6 @@ export default function Page(props) {
       setApiKeys(response.apiKeys);
       setTotal(response.total);
     }
-  }
-
-  function updateURLParams() {
-    let url = `${window.location.pathname}?`;
-    if (offset && offset !== 0) url += `offset=${offset}&`;
-
-    const routerMethod =
-      url === `${window.location.pathname}?` ? "replace" : "push";
-    router[routerMethod](router.route, url, { shallow: true });
   }
 
   return (
