+++ core/web/pages/destinations.tsx
@@ -2,10 +2,9 @@ import Head from "next/head";
 import { Button, Badge } from "react-bootstrap";
 import { useRouter } from "next/router";
 import { useApi } from "../hooks/useApi";
-import { useOffset } from "../hooks/useOffset";
+import { useOffset, updateURLParams } from "../hooks/URLParams";
 import { useState } from "react";
 import { useSecondaryEffect } from "../hooks/useSecondaryEffect";
-import { useHistoryPagination } from "../hooks/useHistoryPagination";
 import Link from "next/link";
 import Pagination from "../components/pagination";
 import LoadingTable from "../components/loadingTable";
@@ -28,14 +27,13 @@ export default function Page(props) {
   // pagination
   const limit = 100;
   const { offset, setOffset } = useOffset();
-  useHistoryPagination(offset, "offset", setOffset);
 
   useSecondaryEffect(() => {
     load();
   }, [offset, limit]);
 
   async function load() {
-    updateURLParams();
+    updateURLParams(router, { offset });
     setLoading(true);
     const response = await execApi("get", `/destinations`, {
       limit,
@@ -50,15 +48,6 @@ export default function Page(props) {
         router.push("/destination/new");
       }
     }
-  }
-
-  async function updateURLParams() {
-    let url = `${window.location.pathname}?`;
-    if (offset && offset !== 0) url += `offset=${offset}&`;
-
-    const routerMethod =
-      url === `${window.location.pathname}?` ? "replace" : "push";
-    router[routerMethod](router.route, url, { shallow: true });
   }
 
   return (
