+++ core/web/pages/files.tsx
@@ -4,9 +4,8 @@ import { useRouter } from "next/router";
 import Link from "next/link";
 import { useState } from "react";
 import { useApi } from "../hooks/useApi";
-import { useOffset } from "../hooks/useOffset";
+import { useOffset, updateURLParams } from "../hooks/URLParams";
 import { useSecondaryEffect } from "../hooks/useSecondaryEffect";
-import { useHistoryPagination } from "../hooks/useHistoryPagination";
 import Moment from "react-moment";
 import Pagination from "../components/pagination";
 import LoadingTable from "../components/loadingTable";
@@ -27,7 +26,6 @@ export default function Page(props) {
   // pagination
   const limit = 100;
   const { offset, setOffset } = useOffset();
-  useHistoryPagination(offset, "offset", setOffset);
   const csrfToken = globalThis?.localStorage?.getItem("session:csrfToken");
 
   useSecondaryEffect(() => {
@@ -35,7 +33,7 @@ export default function Page(props) {
   }, [limit, offset]);
 
   async function load() {
-    updateURLParams();
+    updateURLParams(router, { offset });
     setLoading(true);
     const response = await execApi("get", `/files`, {
       limit,
@@ -64,15 +62,6 @@ export default function Page(props) {
       }
       setLoading(false);
     }
-  }
-
-  function updateURLParams() {
-    let url = `${window.location.pathname}?`;
-    if (offset && offset !== 0) url += `offset=${offset}&`;
-
-    const routerMethod =
-      url === `${window.location.pathname}?` ? "replace" : "push";
-    router[routerMethod](router.route, url, { shallow: true });
   }
 
   return (
