+++ core/web/pages/resque/locks.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { useApi } from "../../hooks/useApi";
-import { useOffset } from "../../hooks/useOffset";
+import { useOffset, updateURLParams } from "../../hooks/URLParams";
 import { Table, Row, Col } from "react-bootstrap";
 import Pagination from "../../components/pagination";
 import { useRouter } from "next/router";
@@ -25,7 +25,7 @@ export default function ResqueLocksList(props) {
   }, [offset, limit]);
 
   async function load() {
-    updateURLParams();
+    updateURLParams(router, { offset });
     setLoading(true);
     const response = await execApi("get", `/resque/locks`);
     const _locks = [];
@@ -47,13 +47,6 @@ export default function ResqueLocksList(props) {
       successHandler.set({ message: "lock deleted" });
       await load();
     }
-  }
-
-  function updateURLParams() {
-    let url = `${window.location.pathname}`;
-    if (offset && offset !== 0) url += `?offset=${offset}&`;
-
-    router.push(router.route, url, { shallow: true });
   }
 
   return (
